fix(logs): split operations and messages index
