+++ packages/logs/lib/es/helpers.ts
@@ -3,9 +3,10 @@ import { isTest } from '@nangohq/utils';
 import { envs } from '../env.js';
 import { logger } from '../utils.js';
 import { client } from './client.js';
-import { getDailyIndexPipeline, indexMessages, policyRetention } from './schema.js';
-import { getFormattedMessage } from '../models/helpers.js';
+import { getDailyIndexPipeline, indexMessages, indexOperations, policyRetention } from './schema.js';
+import { getFormattedMessage, getFormattedOperation } from '../models/helpers.js';
 import { createMessage } from '../models/messages.js';
+import { createOperation } from '../models/operations.js';
 
 export async function start() {
     if (!envs.NANGO_LOGS_ENABLED) {
@@ -22,39 +23,43 @@ export async function start() {
 
 export async function migrateMapping() {
     try {
-        const index = indexMessages;
+        for (const index of [indexMessages, indexOperations]) {
+            logger.info(`Migrating index "${index.index}"...`);
 
-        logger.info(`Migrating index "${index.index}"...`);
+            // -- Policy
+            logger.info(`  Updating policy`);
+            await client.ilm.putLifecycle(policyRetention());
 
-        // -- Policy
-        logger.info(`  Updating policy`);
-        await client.ilm.putLifecycle(policyRetention());
+            // -- Index
+            const existsTemplate = await client.indices.existsIndexTemplate({ name: `${index.index}-template` });
+            logger.info(`  ${existsTemplate ? 'updating' : 'creating'} index template "${index.index}"...`);
 
-        // -- Index
-        const existsTemplate = await client.indices.existsIndexTemplate({ name: `${index.index}-template` });
-        logger.info(`  ${existsTemplate ? 'updating' : 'creating'} index template "${index.index}"...`);
+            await client.indices.putIndexTemplate({
+                name: `${index.index}-template`,
+                index_patterns: `${index.index}.*`,
+                template: {
+                    settings: index.settings!,
+                    mappings: index.mappings!,
+                    aliases: { [index.index]: {} }
+                }
+            });
 
-        await client.indices.putIndexTemplate({
-            name: `${index.index}-template`,
-            index_patterns: `${index.index}.*`,
-            template: {
-                settings: index.settings!,
-                mappings: index.mappings!,
-                aliases: { [index.index]: {} }
+            // -- Pipeline
+            // Pipeline will automatically create an index based on a field
+            // In our case we create a daily index based on "createdAt"
+            logger.info(`  Updating pipeline`);
+            await client.ingest.putPipeline(getDailyIndexPipeline(index.index));
+
+            const existsAlias = await client.indices.exists({ index: index.index });
+            if (!existsAlias) {
+                // insert a dummy record to create first index
+                logger.info(`  Inserting dummy record`);
+                if (index.index.includes('messages')) {
+                    await createMessage(getFormattedMessage({ parentId: '-1', accountId: 0 }));
+                } else {
+                    await createOperation(getFormattedOperation({ id: '-1', accountId: 0, operation: { type: 'sync', action: 'run' } }));
+                }
             }
-        });
-
-        // -- Pipeline
-        // Pipeline will automatically create an index based on a field
-        // In our case we create a daily index based on "createdAt"
-        logger.info(`  Updating pipeline`);
-        await client.ingest.putPipeline(getDailyIndexPipeline(index.index));
-
-        const existsAlias = await client.indices.exists({ index: index.index });
-        if (!existsAlias) {
-            // insert a dummy record to create first index
-            logger.info(`  Inserting dummy record`);
-            await createMessage(getFormattedMessage({ parentId: '-1', accountId: 0 }));
         }
     } catch (err) {
         logger.error(err);
