+++ packages/logs/lib/index.ts
@@ -2,11 +2,12 @@ export { migrateMapping as migrateLogsMapping, start } from './es/helpers.js';
 export * from './client.js';
 export * from './models/helpers.js';
 export * from './models/logContextGetter.js';
-export * as model from './models/messages.js';
-export * as modelOperations from './models/insights.js';
+export * as modelMessages from './models/messages.js';
+export * as modelOperations from './models/operations.js';
+export * as modelInsights from './models/insights.js';
 export * from './otlp/otlp.js';
 export { OtlpSpan } from './otlp/otlpSpan.js';
 export { defaultOperationExpiration, envs } from './env.js';
-export { destroy } from './utils.js';
+export { ResponseError, destroy } from './utils.js';
 export * from './formatters.js';
 export * from './transport.js';
