+++ packages/server/lib/controllers/v1/logs/getOperation.ts
@@ -1,6 +1,6 @@
 import * as z from 'zod';
 
-import { envs, model, operationIdRegex } from '@nangohq/logs';
+import { ResponseError, envs, modelOperations, operationIdRegex } from '@nangohq/logs';
 import { requireEmptyQuery, zodErrorToHTTP } from '@nangohq/utils';
 
 import { asyncWrapper } from '../../../utils/asyncWrapper.js';
@@ -35,7 +35,7 @@ export const getOperation = asyncWrapper<GetOperation>(async (req, res) => {
 
     const { environment, account } = res.locals;
     try {
-        const operation = await model.getOperation({ id: val.data.operationId });
+        const operation = await modelOperations.getOperation({ id: val.data.operationId });
         if (operation.accountId !== account.id || operation.environmentId !== environment.id || !operation.operation) {
             res.status(404).send({ error: { code: 'not_found' } });
             return;
@@ -43,7 +43,7 @@ export const getOperation = asyncWrapper<GetOperation>(async (req, res) => {
 
         res.status(200).send({ data: operation });
     } catch (err) {
-        if (err instanceof model.ResponseError && err.statusCode === 404) {
+        if (err instanceof ResponseError && err.statusCode === 404) {
             res.status(404).send({ error: { code: 'not_found' } });
             return;
         }
